File Uploads
Due to how massive the data set, the upload of the project had a problem. This update uploads both the jupyter notebooks and a text file that takes you to a link to the data set. The data set is a 532 MB zip file uploaded to google drive. Unzip the zip file to the same directory as the jupyter notebooks, and the code will function as normal.